requirements.txt
mysql-connector-python
configparser
requests
zipfile
io
csv
black